[Will Dyson] patches to build with gcc 3.4 and 4.0, on a 64-bit system
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2053 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/ModelView/Dim3_Loader.cpp
@@ -225,7 +225,7 @@ bool LoadModel_Dim3(FileSpecifier& Spec, Model3D& Model, int WhichPass)
 	{
 		size_t NumBones = Model.Bones.size();
 		BoneIndices.resize(NumBones);
-		fill(BoneIndices.begin(),BoneIndices.end(),UNONE);	// No bones listed -- yet
+		fill(BoneIndices.begin(),BoneIndices.end(),(size_t)UNONE);	// No bones listed -- yet
 		vector<Model3D_Bone> SortedBones(NumBones);
 		vector<size_t> BoneStack(NumBones);
 		vector<bool> BonesUsed(NumBones);
